chore: rename module
