+++ packages/backend/webserver/webserver.go
@@ -7,7 +7,7 @@ import (
 	"net"
 	"net/http"
 
-	"github.com/kiel-live/kiel-live/backend/websocket"
+	"github.com/kiel-live/kiel-live/packages/backend/websocket"
 )
 
 type WebServer struct {
