+++ packages/backend/websocket/websocketConnection.go
@@ -8,7 +8,7 @@ import (
 	"sync"
 
 	"github.com/gorilla/websocket"
-	"github.com/kiel-live/kiel-live/backend/proto"
+	"github.com/kiel-live/kiel-live/packages/backend/proto"
 	"github.com/pborman/uuid"
 )
